Modularizing code into separate files
