+++ Camera.py
@@ -1,35 +1,28 @@
-from Elements import Point, Vector
+from Elements import Point, Vector, point_subtract, cross
 
-def pointSubtract(p1: Point, p2: Point):
-    v = Vector(0,0,0)
+# defines camera and raycasting
 
-    v.x = p1.x - p2.x
-    v.y = p1.y - p2.y
-    v.z = p1.z - p2.z
+class Ray:
+    def __init__(self, origin: Point, direction: Vector):
+        self.origin = origin
+        self.direction = direction.get_normalized()
 
-    return v
-def cross(v1: Vector, v2: Vector):
-    product = Vector(v1.y*v2.z - v2.z*v2.y,
-                     v1.z*v2.x - v2.x*v2.z,
-                     v1.x*v2.y - v2.y*v2.x,)
-    return product
+    def __repr__(self):
+        return f"Ray({self.origin}, {self.direction})"
+
+    def __str__(self):
+        return f"Ray({self.origin}, {self.direction})"
 
 class Camera:
-    def __init__(self, pos: Point, target: Point, up: Vector, dist, hScreen, wScreen):
+    def __init__(self, pos: Point, target: Point, up: Vector, dist: float, hScreen: int, wScreen: int):
         self.pos = pos
         self.target = target
-    
-        # Vectors of R3
-        self.w = pointSubtract(pos, target).getNormalized()
-        self.u = cross(self.w, up).getNormalized()
-        self.up = up.getNormalized()
-
+        self.w = point_subtract(pos, target).get_normalized()
+        self.u = cross(up, self.w).get_normalized()
+        self.v = cross(self.w, self.u)
         self.dist = dist
         self.hScreen = hScreen
         self.wScreen = wScreen
 
-    def __repr__(self):  # __repr__ is used for debugging
-        return f"Camera({self.pos}, {self.target}, {self.up})"
-
-    def __str__(self):  # __str__ is used for printing
-        return f"({self.pos}, {self.target}, {self.up})"
+    def __repr__(self):
+        return f"Camera({self.pos}, {self.target}, {self.dist}, {self.hScreen}, {self.wScreen})"
